feat: project refactor to monorepo
