+++ .npmrc
@@ -1,0 +1,1 @@
+node-linker=hoisted
