+++ README.md
@@ -1,5 +1,21 @@
 # OutbreakX
+
 OutbreakX is an open-source platform for infectious disease surveillance, using geospatial data to track outbreaks in real time. Built with ExpresJS, NodeJS, ReactJS, PostgreSQL with PostGIS, and OpenStreetMaps.
 
 ## Welcome to All Contributors
+
 This is an initiation for an opensource platform for infectious disease surveillance with geospacial data. We welcome all opensource contributors not only to build this system with your codes but also improve this system with your ideas.
+
+Start frontend workspace
+
+`pnpm --filter frontend start`
+
+start backend workspace
+
+`pnpm --filter backend dev`
+
+start project using turbo repo
+
+`turbo run dev`
+
+Note: If any error check the turborepo config or individual package.json file.
