+++ apps/backend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "outbreakx",
+  "name": "backend",
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
