+++ apps/frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "outbreak-x",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
