+++ apps/frontend/public/favicon.ico
