+++ apps/frontend/public/logo512.png
