+++ apps/frontend/src/App.css
