+++ apps/frontend/src/components/MapComponent.tsx
